requirements.txt
mysql-connector-python
flask
sqlalchemy
flask-mysqldb